+++ Simulator_VK_GB_/Models/Groups.swift
@@ -8,6 +8,7 @@ import Foundation
 import SwiftyJSON
 
 class Groups {
+    
     var name: String = ""
     var photoUrl: String = ""
     var avatar: URL? { URL(string: "\(photoUrl)") }
